requirements.txt
opencv-python
numpy
pyautogui
pynput
requests
pyyaml
PySide6
ruamel.yaml
pyinstaller
pywin32; sys_platform == "win32"
windows-capture ; sys_platform == "win32"
pyobjc-framework-Quartz ; sys_platform == "darwin" # TODO: dev macOS in another branch